Move CPU kernels used by shared providers into a separate header
